BDT Scenarios - Change the location of a meeting - refactored
